audio-refactoring: rename audio-drivers to driver_$PLATFORM_$DRIVER
